Change IMU interface to do res conversion internally

+++ edisondrone/include/imu.h
@@ -1,8 +1,6 @@
 #ifndef IMU_H
 #define IMU_H
 
-#include <cstdint>
-
 namespace EdisonDrone {
 
     class IMU {
@@ -12,8 +10,8 @@ namespace EdisonDrone {
             virtual void start() = 0;
             virtual void readGyro() = 0;
             virtual void readAccel() = 0;
-            virtual void getGyroVals(int16_t *x, int16_t *y, int16_t *z) = 0;
-            virtual void getAccelVals(int16_t *x, int16_t *y, int16_t *z) = 0;
+            virtual void getGyroVals(double *x, double *y, double *z) = 0;
+            virtual void getAccelVals(double *x, double *y, double *z) = 0;
 
     };
